Pull out the appointment time picker
